% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_check}
\alias{add_check}
\title{Add a check to a canonical form}
\usage{
add_check(cf, ...)
}
\arguments{
\item{cf}{CanonicalForm object}

\item{...}{named arguments where each name is the name of the check, and each value
is a check function}
}
\value{
a CanonicalForm object with the additional check
}
\description{
Add a check to a canonical form
}
