% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{add_checks}
\alias{add_checks}
\alias{delete_check}
\title{Add checks or delete check}
\usage{
add_checks(cf, ...)

delete_check(cf, check_name)
}
\arguments{
\item{cf}{a CanonicalForm object}

\item{...}{named arguments. Each name will be the name of the check, and
each value should be a check function}

\item{check_name}{the name of the check to delete}
}
\value{
modified CanonicalForm object
}
\description{
Note that this does not work in place. See examples
}
\section{Functions}{
\itemize{
\item \code{delete_check}: Delete a check
}}

\examples{
cf <- extract_canonical_form(cars)
check <- function(x) TRUE
cf <- add_checks(cf, new_check = check) # note this does not work in place
print(get_check_names(cf))
cf <- delete_check(cf, "new_check")
print(get_check_names(cf))
}
