% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{apply_comparisons_from_named_list}
\alias{apply_comparisons_from_named_list}
\title{Apply comparisons from a named list}
\usage{
apply_comparisons_from_named_list(x, comparison, compare_to)
}
\arguments{
\item{x}{data.frame or similar}

\item{comparison}{a comparison function}

\item{compare_to}{a named list of the form list(colname = known_values).}
}
\value{
a named logical vector
}
\description{
Apply a comparison function to multiple columns of a data frame,
which different known values specified for each column.
}
\examples{

x <- data.frame(a = 1:2, b = 3:4)
comparison <- function(e1, e2) any(e1 == e2)
#' # returns c(a = TRUE, b = FALSE)
apply_comparisons_from_named_list(x, comparison, list(a = 1, b = 1))
}
\keyword{internal}
