% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{apply_function_to_cols}
\alias{apply_function_to_cols}
\title{Internal: apply a function to columns of a dataset,}
\usage{
apply_function_to_cols(x, f, function_args, agg = c("all", "any", "none"))
}
\arguments{
\item{function_args}{a named list, where the names are columns of .x
and the values are arguments to the function}

\item{.x}{a data.frame-like object}

\item{.f}{a function, probably with arguments representing (actual, expected)}
}
\value{
a named list with the same names as function_args
}
\description{
With different arguments to each column. This is used by many of the
check functions.
}
