% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{canonical_form}
\alias{canonical_form}
\title{Title}
\usage{
canonical_form(col_names, col_classes, transformers = list(), checks = list())
}
\arguments{
\item{transformers}{}
}
\value{

}
\description{
Title
}
