% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{canonical_object_class}
\alias{canonical_object_class}
\alias{canonical_col_names}
\alias{canonical_col_classes}
\title{Access the properties of a CanonicalForm}
\usage{
canonical_object_class(cf = NULL)

canonical_col_names(cf = NULL)

canonical_col_classes(cf = NULL)
}
\arguments{
\item{cf}{a CanonicalForm object}
}
\value{
The value of a property
}
\description{
These functions allow you to access the properties of a CanonicalForm
from inside of another function. See examples.
}
\section{Functions}{
\itemize{
\item \code{canonical_col_names}: Canonical object class

\item \code{canonical_col_classes}: Canonical column classes
}}

\examples{
check_col_names_in_canonical <- function(x) all(colnames(x) \%in\% canonical_col_names())
cf <- extract_canonical_form(cars)
run_check(check_col_names_in_canonical, cf, x = cars)
}
