% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanonicalForms-package.R
\docType{package}
\name{CanonicalForms-package}
\alias{CanonicalForms}
\alias{CanonicalForms-package}
\title{CanonicalForms: Check that datasets are canonical.}
\description{
Easily specify a "Canonical Form" for a dataset. A CanonicalForm object is a simple and lightweight container for properties and data validation checks that can be applied to a dataset. This package is designed to be minimally invasive to existing workflows, so that data validation checks can be applied with just one or two simple function calls that fit seamlessly with existing data processing steps.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/CanonicalForms}
  \item Report bugs at \url{https://github.com/colin-fraser/CanonicalForms/issues}
}

}
\author{
\strong{Maintainer}: Colin Fraser \email{colin.r.fraser@gmail.com}

}
\keyword{internal}
