% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_class}
\alias{check_class}
\alias{check_col_names}
\alias{check_col_classes}
\alias{default_checks}
\title{Built-in checks}
\usage{
check_class(x)

check_col_names(x)

check_col_classes(x)

default_checks()
}
\arguments{
\item{x}{a dataframe to check}
}
\description{
These checks are built-in
}
\section{Functions}{
\itemize{
\item \code{check_col_names}: check that the column names match the Canonical Form

\item \code{check_col_classes}: check that the classes match the Canonical Form

\item \code{default_checks}: returns a list of the default checks.
}}

