% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_factor_levels}
\alias{check_factor_levels}
\title{Check factor levels}
\usage{
check_factor_levels(..., .order_matters = TRUE)
}
\arguments{
\item{...}{named list where each name is a column name
and each value is the expected factor vector}
}
\value{
a check result which passes if each column's levels
  are equal to the given canonical levels
}
\description{
Check factor levels
}
\examples{
df <- data.frame(
  a = as.factor(c("a", "b", "c", "c")),
  b = as.factor(c("a", "b", "c", "d"))
)
passer <- check_factor_levels(a = c("a", "b", "c"), b = c("a", "b", "c", "d"))
failer <- check_factor_levels(a = c("a", "b", "c"), b = c("d", "e", "f", "g"))
passer(df)
failer(df)
passer2 <- check_factor_levels(a = c('b', 'c', 'a'),
                               b = c('b', 'c', 'a', 'd'),
                               .order_matters = FALSE)
failer2 <- check_factor_levels(a = c('b', 'c', 'a'),
                               b = c('b', 'c', 'a', 'd'),
                               .order_matters = TRUE)
passer2(df)
failer2(df)
}
