% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_no_nas}
\alias{check_no_nas}
\alias{check_greater_than}
\alias{check_less_than}
\title{Check makers}
\usage{
check_no_nas(cols)

check_greater_than(..., .strict = TRUE)

check_less_than(..., .strict = TRUE)
}
\arguments{
\item{cols}{character vector of columns to include in the check}

\item{...}{named arguments of the form (col = constraint)}

\item{.strict}{should values equal to the constraint be disallowed?}
}
\description{
Create checks to add to Canonical Forms. Each of these functions returns a function which takes
a dataframe and returns a CheckResult object.
}
\details{
for \code{check_greater_than} and \code{check_less_than}, different values can be specified for each
   column. See examples.
}
\section{Functions}{
\itemize{
\item \code{check_greater_than}: check that all values are above specified values

\item \code{check_less_than}: check all values are less than constraints
}}

\examples{
df <- data.frame(a = c(1, 2, NA), b = c(4, 5, 6))
check_na <- check_no_nas("a")
check_na(df) # fails
check_na2 <- check_no_nas("b")
check_na2(df) # passes

check_gt <- check_greater_than(a = 0, b = 3)
check_gt(df)  # passes
check_gt2 = check_greater_than(a = 1, b = 4, .strict = FALSE)
check_gt2(df)  # passes (because .strict=FALSE)
check_gt3 <- check_greater_than(a = 1, b = 1, .strict = TRUE)
check_gt3(df)  # fails
}
