% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{default_checks}
\alias{default_checks}
\title{Checks added to canonical forms by default}
\usage{
default_checks(form)
}
\arguments{
\item{form}{a CanonicalForm}
}
\description{
Checks added to canonical forms by default
}
