% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{format.CanonicalForm}
\alias{format.CanonicalForm}
\title{Format a canonical form}
\usage{
\method{format}{CanonicalForm}(x, ...)
}
\arguments{
\item{x}{a CanonicalForm}

\item{...}{other arguments passed to format}
}
\description{
Format a canonical form
}
