% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{get_check_names}
\alias{get_check_names}
\title{List the names of checks associated with a CanonicalForm}
\usage{
get_check_names(cf)
}
\arguments{
\item{cf}{a CanonicalForm object}
}
\value{
character vector of names of checks
}
\description{
List the names of checks associated with a CanonicalForm
}
\examples{
cf <- extract_canonical_form(cars)
get_check_names(cf)
}
