% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{is_canonical}
\alias{is_canonical}
\title{Check if a dataset is canonical}
\usage{
is_canonical(x, form, verbose = TRUE)
}
\arguments{
\item{x}{a dataset}

\item{form}{a CanonicalForm}

\item{verbose}{print check results?}
}
\value{
logical value indicating if the dataset is canonical
}
\description{
Check if a dataset is canonical
}
