% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{make_check_no_nas}
\alias{make_check_no_nas}
\title{Built-in check makers}
\usage{
make_check_no_nas(cols)
}
\arguments{
\item{cols}{columns to apply the check to}
}
\value{
a function that can be added to canonical form checks
}
\description{
Built-in check makers
}
