% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{named_logical_vector_to_check_result}
\alias{named_logical_vector_to_check_result}
\title{Create a CheckResult from a named logical vector}
\usage{
named_logical_vector_to_check_result(vec, header)
}
\arguments{
\item{vec}{a named logical vector}

\item{header}{the header for the message that will be created.}
}
\value{
a CheckResult object
}
\description{
This takes a named logical vector and turns it into a CheckResult
object, using the names to create a message.
}
\examples{
vec <- c(a = TRUE, b = FALSE)
# test fails, and message includes "b" as the reason for failing
res <- named_logical_vector_to_check_result(vec, "A test!")
print(res)
#> failed check
#>
#> Additional info:
#>   A test!
#>   x b
}
