% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{print.CanonicalForm}
\alias{print.CanonicalForm}
\title{Print Canonical Form}
\usage{
\method{print}{CanonicalForm}(x, ...)
}
\arguments{
\item{x}{a CanonicalForm}

\item{...}{other arguments}
}
\description{
Print Canonical Form
}
