% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_all_checks}
\alias{run_all_checks}
\title{Run all checks}
\usage{
run_all_checks(x, cf)
}
\arguments{
\item{x}{the object to check against, probably a data.frame or tibble}

\item{cf}{a canonical_form}
}
\value{
a list of check_result objects
}
\description{
Run all the tests registered to CanonicalForm
}
\examples{
cf <- extract_canonical_form(cars)
run_all_checks(cars, cf)
}
