% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{run_check}
\alias{run_check}
\title{Run a check}
\usage{
run_check(check, cf, ...)
}
\arguments{
\item{check}{a function to run}

\item{cf}{a CanonicalForm object}

\item{...}{additional arguments passed to the check function}
}
\value{
the return value of check
}
\description{
This is the primary way that checks are run programatically.
This runs the check in the \code{check_env} environment, which
allows it to access properties about the CanonicalForm such as
canonical_col_names().
}
\examples{
cf <- extract_canonical_form(cars)
run_check("check_class", cf, x = cars)
}
