% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{to_r_code}
\alias{to_r_code}
\title{Get the R code to create a CanonicalForm}
\usage{
to_r_code(x)
}
\arguments{
\item{x}{a canonical form}
}
\value{

}
\description{
Get the R code to create a CanonicalForm
}
