% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{canonical_form}
\alias{canonical_form}
\title{Create a CanonicalForm}
\usage{
canonical_form(
  object_class,
  col_names,
  col_classes,
  transformers = list(),
  checks = list(),
  add_default_checks = TRUE
)
}
\arguments{
\item{col_names}{a vector of column names}

\item{col_classes}{a vector of column classes}

\item{transformers}{a list of canonicalization functions}

\item{checks}{a list of functions that will be used to check
if a dataset is canonical}
}
\value{
an object of class CanonicalForm
}
\description{
Create a CanonicalForm
}
