% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_forms.R
\name{canonical_form}
\alias{canonical_form}
\title{Create a CanonicalForm}
\usage{
canonical_form(
  object_class,
  col_names,
  col_classes,
  checks = list(),
  uncategorized_checks = list(),
  add_default_checks = TRUE
)
}
\arguments{
\item{object_class}{the class of the object}

\item{col_names}{a vector of column names}

\item{col_classes}{a vector of column classes}

\item{uncategorized_checks}{a list of checks}

\item{add_default_checks}{should the default checks be added?}

\item{transformers}{a list of canonicalization functions}

\item{global_checks}{a list of functions that will be used to check
if a dataset is canonical}

\item{column_checks}{a list of checks}
}
\value{
an object of class CanonicalForm
}
\description{
Create a CanonicalForm
}
