% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CanonicalForms-package.R
\docType{package}
\name{CanonicalForms-package}
\alias{CanonicalForms}
\alias{CanonicalForms-package}
\title{CanonicalForms:}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/colin-fraser/CanonicalForms}
  \item Report bugs at \url{https://github.com/colin-fraser/CanonicalForms/issues}
}

}
\keyword{internal}
