% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{checks}
\alias{checks}
\alias{check_class}
\alias{check_colnames}
\alias{check_col_classes}
\title{Built-in checks}
\usage{
check_class(form)

check_colnames(form)

check_col_classes(form)
}
\arguments{
\item{form}{canonical form to check against}
}
\value{
logical indicating whether the check is passed
}
\description{
Built-in checks
}
